Routing means navigation between pages. It enables developers to build a Single Page Application (SPA) with multiple views and allow navigation between views.
Whenever we creating an angular project, the app-routing.modules.ts file is automatically created, which is a core file of routing to handle all the routes.

Router-outlet is used to load the different components based on the activated route. Router-outlet is a directive that is available in the router library. You can add the multiple router-outlet in your application.